Add files via upload
This script is responsible for training the SMS Spam-Ham detection model.
It performs data loading, preprocessing, feature extraction (BoW / TF-IDF),
and trains a Logistic Regression classifier.  
The trained model and vectorizer are saved as `spam_detection.pkl` and BoW_vectorizer.pkl for later use 
in the Flask API or other applications.